The package 'exceptions' has been moved to an external location.
